requirements.txt
PyYAML==6.0.2
requests==2.32.5
watchdog==6.0.0
fastapi==0.116.1
uvicorn==0.35.0
ruamel.yaml==0.18.16
